add detailPage,loading,404

+++ src/store/slices/userSlice.js
@@ -1,6 +1,7 @@
 import { createSlice, createAsyncThunk } from "@reduxjs/toolkit";
 import { login } from "../../services/user";
-
+import Swal from "sweetalert2";
+import routes from "../../routes/routes";
 //redux saga + react reduc : redux tookit 없이 쓰는 법. 툴킷 안쓰는 회사도 많음
 
 const initialState = {
@@ -47,24 +48,26 @@ export const loginRequest = createAsyncThunk(
         const response = await login({ email, password });
 
         if (response.status === 200) {
-            alert('로그인성공')
-                .then(() => {
-
+            Swal.fire({
+                icon: 'success',
+                title: '로그인 성공!',
+                text: '홈 화면으로 이동합니다',
+                confirmButtonColor: '#429f50',
+            }).then(result => {
+                if (result.isConfirmed) {
                     localStorage.setItem("email", email);
                     localStorage.setItem("token", response.headers.authorization);
-
-                })
-                .then(() => {
-                    window.location.href = "/home";
-                })
-
-            return {
-                email: email,
-                token: response.headers.authorization,
-            };
-
+                    window.location.href = routes.home;
+                }
+            })
         }
-    })
+        return {
+            email: email,
+            token: response.headers.authorization,
+        };
+
+    }
+)
 //로그인에 대한 비동기 요청을 컴포넌트에서 진행했는데 , 여기서 비동기 
 
 export const { setEmail } = userSlice.actions;
